Implemented distortion map
git-svn-id: http://yuvtoolkit.googlecode.com/svn@147 fe7d2ec1-bd83-e687-9167-72ef5a783f6b

+++ trunk/Plugins/YT_Interface.h
@@ -329,8 +329,8 @@ struct MeasureInfo
 {
 	QString name;
 	QString unit;
-	double upperRange;
-	double lowerRange;
+	float upperRange;
+	float lowerRange;
 	bool biggerValueIsBetter;
 };
 
@@ -342,7 +342,7 @@ struct MeasureCapabilities
 	bool hasColorDistortionMap; // Can generate distortion map for Y, U, or V planes combined
 };
 
-typedef QVector<double> DistMap;
+typedef QVector<float> DistMap;
 typedef QSharedPointer<DistMap> DistMapPtr;
 
 struct MeasureOperation
